+++ src/components/FloatingButton.tsx
@@ -1,17 +1,9 @@
-import { useLocation } from "react-router-dom";
-import BaseIconButton from "../components/common/BaseIconButton";
-import {
-  FloatingDetailData,
-  FloatingHomeData,
-  FloatingIcon,
-  projectLinks,
-} from "../data/FloatingData";
 import React from "react";
+import { useLocation } from "react-router-dom";
+import BaseIconButton from "./common/BaseIconButton";
+import { FloatingDetailData, FloatingHomeData, FloatingIcon, } from "../data/FloatingData";
+import { ReactIconProps } from "../utils/types";
 
-type IconProps = {
-  size?: number;
-  className?: string;
-};
 
 const FloatingButton = () => {
   const { pathname } = useLocation();
@@ -27,17 +19,17 @@ const FloatingButton = () => {
     `}
     >
       {floatingData.map(({ key, value, title, type }) => {
-        const Icon = FloatingIcon[key] as React.ComponentType<IconProps>;
+        const Icon = FloatingIcon[key] as React.ComponentType<ReactIconProps>;
         if (!Icon) return null;
 
         const props =
           type === "scroll"
             ? { onClick: () => window.scrollTo({ top: 0, behavior: "smooth" }) }
             : {
-                href: `https://${value}`,
-                target: "_blank",
-                rel: "noopener noreferrer",
-              };
+              href: `https://${value}`,
+              target: "_blank",
+              rel: "noopener noreferrer",
+            };
 
         return (
           <BaseIconButton key={key} title={title} {...props}>
